Fixed bug of "resource busy" message received.
Have to wait more after calling StartBMode,
probably for the buffers to flush from previous data.

Changed log level of libusb to none.
Now, the multithread asynchronous gets data back everytime,
but there is still the problem of glitches that need to be fixed.

Also changed to shared libusb library that the program uses,
to one that I compiled instead of the debian pkg one.
it fixed the problem of the program having a segfault when libusb
returns "Resource Busy". There is probably a bug in the dpkg version.